Struct hid_io::protocol::hidio::HIDIOPacketBuffer

pub struct HIDIOPacketBuffer {
    pub ptype: HIDIOPacketType,
    pub id: u32,
    pub max_len: u32,
    pub data: Vec<u8>,
    pub done: bool,
}

HID-IO Packet Buffer Struct

Remarks

Used to store HID-IO data chunks. Will be chunked into individual packets on transmission.

Fields
ptype: HIDIOPacketType

Type of packet (Continued is automatically set if needed)

id: u32

Packet Id

max_len: u32

Packet length for serialization (in bytes)

data: Vec<u8>

Payload data, chunking is done automatically by serializer

done: bool

Set False if buffer is not complete, True if it is

Methods

impl HIDIOPacketBuffer

pub fn new() -> HIDIOPacketBuffer

Constructor for HIDIOPacketBuffer

Remarks

Initialize as blank

pub fn append_payload(&mut self, new_data: &mut Vec<u8>) -> bool

Append payload data

Arguments

new_data - Vector of bytes

Remarks

Appends payload to HIDIOPacketBuffer.

pub fn decode_packet(
&mut self,
packet_data: &mut Vec<u8>
) -> Result<u32, HIDIOParseError>

Append packet stream Returns the number of bytes used.

Arguments

packet_data - Vector of bytes of packet data

Remarks

Does packet decoding on the fly. Will set done parameter if this is the last packet.

pub fn serialize_buffer(&mut self) -> Result<Vec<u8>, HIDIOParseError>

Serialize HIDIOPacketBuffer

Remarks

Provides a raw data vector to the serialized data. Removes some of the header that Serialize from serde prepends.